refactor to move payment processing into it’s own effect
- this is the first step in trying to improve the createSource behaviour.
